CHE-747: Split GWT compilation and wsmaster API war. Add module wsmaster Move ProjectTemplateService and all classes around to separate module (ws master part) Rename assembly modules *-machines-* to the *-wsagent-* Fix rewrite rulls accordint to rename war file (dashboard requre it)
