+++ assembly/assembly-main/pom.xml
@@ -29,8 +29,13 @@
         </dependency>
         <dependency>
             <groupId>org.eclipse.che</groupId>
-            <artifactId>assembly-machine-server</artifactId>
+            <artifactId>assembly-wsagent-server</artifactId>
             <type>zip</type>
+        </dependency>
+        <dependency>
+            <groupId>org.eclipse.che</groupId>
+            <artifactId>assembly-wsmaster-war</artifactId>
+            <type>war</type>
         </dependency>
         <dependency>
             <groupId>org.eclipse.che.core</groupId>
