+++ assembly/assembly-main/src/assembly/webapps/ROOT/WEB-INF/rewrite.config
@@ -1,3 +1,3 @@
 
-RewriteRule ^/api/(.*)$     /ide/api/$1    [L]
+RewriteRule ^/api/(.*)$     /wsmaster/api/$1    [L]
 RewriteRule ^/$    /dashboard    [R]
