+++ assembly/assembly-wsagent-server/src/assembly/LICENSE-mit.txt
