+++ assembly/assembly-wsagent-server/src/assembly/setenv.sh
