+++ assembly/assembly-wsagent-war/src/main/java/org/eclipse/che/wsagent/server/WsAgentServletModule.java
