+++ assembly/assembly-wsagent-war/src/main/webapp/WEB-INF/web.xml
