+++ assembly/assembly-wsmaster-war/src/main/resources/predefined-recipes.json
