+++ assembly/assembly-wsmaster-war/src/main/webapp/META-INF/context.xml
