+++ assembly/assembly-wsmaster-war/src/main/webapp/WEB-INF/classes/codenvy/che.properties
@@ -47,7 +47,7 @@ oauth.github.clientid=NULL
 oauth.github.clientsecret=NULL
 oauth.github.authuri= https://github.com/login/oauth/authorize
 oauth.github.tokenuri= https://github.com/login/oauth/access_token
-oauth.github.redirecturis= http://localhost:${SERVER_PORT}/ide/api/oauth/callback
+oauth.github.redirecturis= http://localhost:${SERVER_PORT}/wsmaster/api/oauth/callback
 auth.oauth.access_denied_error_page=/error-oauth
 
 # Activates user self-service at the API level.
@@ -102,7 +102,7 @@ machine.docker.machine_extra_hosts=NULL
 # This is the API endpoint of the workspace master running within the core Che server.
 # This tells the workspace agent how to connect back to the Che server.
 # che-host is a fake DNS hostname that Che creates for itself.
-machine.docker.che_api.endpoint=http://che-host:${SERVER_PORT}/ide/api
+machine.docker.che_api.endpoint=http://che-host:${SERVER_PORT}/wsmaster/api
 
 # If this is true, then we always pull an image from a registry even if we have an image cached
 # locally. If false, Docker only pulls image if it does not exist locally.
@@ -119,7 +119,7 @@ machine.docker.local_node_host=NULL
 
 # URL path to api service.  
 # Browser clients use this to initiate REST communications with workspace master
-api.endpoint=http://localhost:${SERVER_PORT}/ide/api
+api.endpoint=http://localhost:${SERVER_PORT}/wsmaster/api
 
 #### Everrest is a Java Web Services toolkit that manages JAX-RS & web socket communications
 #### Users should rarely need to configure this.
