+++ assembly/pom.xml
@@ -25,9 +25,10 @@
     <packaging>pom</packaging>
     <name>Che IDE :: Parent</name>
     <modules>
-        <module>assembly-machine-war</module>
-        <module>assembly-machine-server</module>
+        <module>assembly-wsagent-war</module>
+        <module>assembly-wsagent-server</module>
         <module>assembly-ide-war</module>
+        <module>assembly-wsmaster-war</module>
         <module>assembly-main</module>
     </modules>
 </project>
