+++ core/ide/che-core-ide-app/src/main/java/org/eclipse/che/ide/projecttype/ProjectTemplatesComponent.java
@@ -11,8 +11,8 @@
 package org.eclipse.che.ide.projecttype;
 
 import org.eclipse.che.api.project.gwt.client.ProjectTemplateServiceClient;
-import org.eclipse.che.api.project.shared.dto.ProjectTemplateDescriptor;
 
+import org.eclipse.che.api.project.templates.shared.dto.ProjectTemplateDescriptor;
 import org.eclipse.che.ide.api.project.type.ProjectTemplateRegistry;
 import org.eclipse.che.ide.api.component.Component;
 import org.eclipse.che.ide.rest.AsyncRequestCallback;
