+++ core/platform-api-client-gwt/che-core-client-gwt-project/src/main/resources/org/eclipse/che/api/Project.gwt.xml
@@ -20,5 +20,6 @@
     <inherits name="org.eclipse.che.api.Workspace"/>
     <source path='project/gwt'/>
     <source path='project/shared'/>
+    <source path='project/templates'/>
     <source path='vfs/impl/file'/>
 </module>
